backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
pydantic==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
requests==2.32.5

# Development and Linting Tools
flake8==7.0.0
black==24.1.1
isort==5.13.2
mypy==1.8.0
pre-commit==3.6.0
pytest==8.0.0
pytest-cov==4.1.0